requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black
black==20.8b1             # via -r requirements.in
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black, pip-tools
idna==2.10                # via requests
mypy-extensions==0.4.3    # via black
pathspec==0.8.0           # via black
pip-tools==5.3.1          # via -r requirements.in
pyyaml==5.3.1             # via -r requirements.in
regex==2020.7.14          # via black
requests==2.24.0          # via todoist-python
selenium==3.141.0         # via -r requirements.in
six==1.15.0               # via pip-tools
todoist-python==8.1.2     # via -r requirements.in
toml==0.10.1              # via black
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
typing==3.7.4.3           # via todoist-python
urllib3==1.25.10          # via requests, selenium

# The following packages are considered to be unsafe in a requirements file:
# pip
